// Qanat.Web/src/app/shared/generated/model/allocation-plan-manage-dto.ts
/**
 * Qanat.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AllocationPlanPeriodSimpleDto } from '././allocation-plan-period-simple-dto';
import { WaterTypeSimpleDto } from '././water-type-simple-dto';
import { ZoneMinimalDto } from '././zone-minimal-dto';
import { GeographyAllocationPlanConfigurationSimpleDto } from '././geography-allocation-plan-configuration-simple-dto';

import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class AllocationPlanManageDto { 
    AllocationPlanID?: number;
    GeographyAllocationPlanConfiguration?: GeographyAllocationPlanConfigurationSimpleDto;
    Zone?: ZoneMinimalDto;
    WaterType?: WaterTypeSimpleDto;
    AllocationPlanPeriods?: Array<AllocationPlanPeriodSimpleDto>;
    LastUpdated?: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface AllocationPlanManageDtoForm { 
    AllocationPlanID?: FormControl<number>;
    GeographyAllocationPlanConfiguration?: FormControl<GeographyAllocationPlanConfigurationSimpleDto>;
    Zone?: FormControl<ZoneMinimalDto>;
    WaterType?: FormControl<WaterTypeSimpleDto>;
    AllocationPlanPeriods?: FormControl<Array<AllocationPlanPeriodSimpleDto>>;
    LastUpdated?: FormControl<string>;
}

export class AllocationPlanManageDtoFormControls { 
    public static AllocationPlanID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static GeographyAllocationPlanConfiguration = (value: FormControlState<GeographyAllocationPlanConfigurationSimpleDto> | GeographyAllocationPlanConfigurationSimpleDto = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<GeographyAllocationPlanConfigurationSimpleDto>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Zone = (value: FormControlState<ZoneMinimalDto> | ZoneMinimalDto = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<ZoneMinimalDto>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static WaterType = (value: FormControlState<WaterTypeSimpleDto> | WaterTypeSimpleDto = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<WaterTypeSimpleDto>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static AllocationPlanPeriods = (value: FormControlState<Array<AllocationPlanPeriodSimpleDto>> | Array<AllocationPlanPeriodSimpleDto> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<AllocationPlanPeriodSimpleDto>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static LastUpdated = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}