// Qanat.Web/src/app/shared/generated/model/allocation-plan-period-upsert-dto.ts
export class AllocationPlanPeriodUpsertDto { 
    AllocationPlanID: number;
    AllocationPlanPeriodID?: number;
    AllocationPeriodName: string;
    AllocationAcreFeetPerAcre: number;
    NumberOfYears: number;
    StartYear: number;
    EnableCarryOver: boolean;
    CarryOverNumberOfYears?: number | null;
    CarryOverDepreciationRate?: number | null;
    EnableBorrowForward: boolean;
    BorrowForwardNumberOfYears?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface AllocationPlanPeriodUpsertDtoForm { 
    AllocationPlanID: FormControl<number>;
    AllocationPlanPeriodID?: FormControl<number>;
    AllocationPeriodName: FormControl<string>;
    AllocationAcreFeetPerAcre: FormControl<number>;
    NumberOfYears: FormControl<number>;
    StartYear: FormControl<number>;
    EnableCarryOver: FormControl<boolean>;
    CarryOverNumberOfYears?: FormControl<number>;
    CarryOverDepreciationRate?: FormControl<number>;
    EnableBorrowForward: FormControl<boolean>;
    BorrowForwardNumberOfYears?: FormControl<number>;
}

export class AllocationPlanPeriodUpsertDtoFormControls { 
    public static AllocationPlanID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static AllocationPlanPeriodID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static AllocationPeriodName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.minLength(1),
            ],
        }
    );
    public static AllocationAcreFeetPerAcre = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.min(0),
                Validators.max(1000),
            ],
        }
    );
    public static NumberOfYears = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
                Validators.min(1),
                Validators.max(100),
            ],
        }
    );
    public static StartYear = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static EnableCarryOver = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static CarryOverNumberOfYears = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.min(1),
                Validators.max(100),
            ],
        }
    );
    public static CarryOverDepreciationRate = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.min(0),
                Validators.max(1),
            ],
        }
    );
    public static EnableBorrowForward = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static BorrowForwardNumberOfYears = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
                Validators.min(1),
                Validators.max(100),
            ],
        }
    );
}